Modification FOMEGA pour correspondre aux differents cas

+++ 5_Resol_Post-Trait/solex.c
@@ -10,21 +10,21 @@ extern int nucas;
 
 float solex(float *coor) {
   const float PI = M_PI;
-  float val = 0.;
+  float val = 0.0f;
 
   switch (nucas) {
     case 1 :
-	val=16.*coor[0]*coor[1]*(coor[0]-1)*(1-coor[1]);
-	break;
+	    val=16.0f*coor[0]*coor[1]*(coor[0]-1)*(1.0f-coor[1]);
+	    break;
     case 2 :
-	val=sin(PI*coor[0])*sin(PI*coor[1]);
-	break;
+	    val=sinf(PI*coor[0])*sinf(PI*coor[1]);
+	    break;
     case 3 :
-	val=cos(PI*coor[0])*cos(PI*coor[1]);
-	break;
+	    val=cosf(PI*coor[0])*cosf(PI*coor[1]);
+	    break;
     default :
-	printf("*** SOLEX : exemple non prevu.\n");
-	break;
+	    printf("*** SOLEX : exemple non prevu.\n");
+	    break;
   }
   return(val);
 }
